Create bridge in stand-alone mode
When launching the CarPlay scene directly on
the CarPlay-client, create a bridge for the AppDelegate.
Again check for an already existing bridge in order to
not create multiple bridges:
If the app is already running on the phone, a bridge will
already be present on the AppDelegate.
If not: create one for the CarPlay scene.